[step3] Rename packages
