+++ src/main/kotlin/week1/calculator/Calculator.kt
@@ -1,4 +1,4 @@
-package week1.step2
+package week1.calculator
 
 import java.util.LinkedList
